+++ components/header/Banner.tsx
@@ -1,6 +1,15 @@
 import { XMarkIcon } from "@heroicons/react/20/solid"
+import getOpenIssues from "@/components/header/utils"
+import { useEffect } from "react"
 
-export default function Banner() {
+export default async function Banner() {
+  const repositories = await getOpenIssues()
+  console.log(repositories)
+  const issues = repositories.filter(
+    (repository) => repository.openIssues.length > 0
+  )
+  console.log(issues)
+
   return (
     <div className="relative isolate flex items-center gap-x-6 overflow-hidden bg-gray-50 px-6 py-2.5 sm:px-3.5 sm:before:flex-1">
       <div
@@ -27,26 +36,32 @@ export default function Banner() {
           className="aspect-[577/310] w-[36.0625rem] bg-gradient-to-r from-[#ff80b5] to-[#9089fc] opacity-30"
         />
       </div>
-      <div className="flex flex-wrap items-center gap-x-4 gap-y-2">
-        <p className="text-sm/6 text-gray-900">
-          <strong className="font-semibold">CCV Services Status</strong>
-          <svg
-            viewBox="0 0 2 2"
-            aria-hidden="true"
-            className="mx-2 inline size-0.5 fill-current"
+      {issues.length > 0 && (
+        <div className="flex flex-wrap items-center gap-x-4 gap-y-2">
+          <p className="text-sm/6 text-gray-900">
+            <strong className="font-semibold">CCV Services Status</strong>
+            <svg
+              viewBox="0 0 2 2"
+              aria-hidden="true"
+              className="mx-2 inline size-0.5 fill-current"
+            >
+              <circle r={1} cx={1} cy={1} />
+            </svg>
+            Service Disruption:
+            {issues.map((repo) => (
+              <span key={repo.name}>{repo.name}</span>
+            ))}
+          </p>
+
+          <a
+            href="https://status.ccv.brown.edu/"
+            target="_blank"
+            className="flex-none rounded-full bg-gray-900 px-3.5 py-1 text-sm font-semibold text-white shadow-sm hover:bg-gray-700 focus-visible:outline focus-visible:outline-2 focus-visible:outline-offset-2 focus-visible:outline-gray-900"
           >
-            <circle r={1} cx={1} cy={1} />
-          </svg>
-          Service Disruption hurrdurr
-        </p>
-        <a
-          href="https://status.ccv.brown.edu/"
-          target="_blank"
-          className="flex-none rounded-full bg-gray-900 px-3.5 py-1 text-sm font-semibold text-white shadow-sm hover:bg-gray-700 focus-visible:outline focus-visible:outline-2 focus-visible:outline-offset-2 focus-visible:outline-gray-900"
-        >
-          View Issues
-        </a>
-      </div>
+            View Issues
+          </a>
+        </div>
+      )}
       <div className="flex flex-1 justify-end">
         <button
           type="button"
