+++ components/header/Header.tsx
@@ -1,5 +1,5 @@
-import Banner from "@/components/Banner"
-import Navbar from "@/components/Navbar"
+import Banner from "@/components/header/Banner"
+import Navbar from "@/components/header/Navbar"
 
 export default function Header() {
   return (
